:art: Refactor travel rates page into a bunch of components.
Why? Easy of modification.
