+++ web/src/components/travel-allowances/NonCommercialAccommodationTravelAllowanceCard.vue
@@ -1,0 +1,20 @@
+<template>
+  <v-card>
+    <v-card-title><h3>Private non-commercial accommodation allowance</h3></v-card-title>
+    <v-card-text>
+      <ul>
+        <li><p>$50 CAD per night (In Yukon, NWT and rest of Canada)</p></li>
+        <li><p>$50 USD per night (In Alaska and rest of USA)</p></li>
+      </ul>
+      <p class="mt-4">
+        Refer to
+        <a
+          href="https://www.njc-cnm.gc.ca/directive/app_d.php?lang=en"
+          target="_blank"
+          >Canada's National Joint Council travel directive, Appendix D</a
+        >.
+      </p>
+      <p></p
+    ></v-card-text>
+  </v-card>
+</template>
