+++ web/src/components/travel-allowances/PrivateVehicleTravelAllowanceCard.vue
@@ -1,0 +1,18 @@
+<template>
+  <v-card>
+    <v-card-title><h3>Distance allowance for private vehicle</h3></v-card-title>
+    <v-card-text>
+      <ul>
+        <li><p>70.0 cents (CAD) per kilometer;</p></li>
+      </ul>
+      <p class="mt-4">
+        Refer to
+        <a
+          href="https://www.njc-cnm.gc.ca/directive/d10/v238/s658/en"
+          target="_blank"
+          >Canada's National Joint Council travel directive, Appendix B</a
+        >.
+      </p>
+    </v-card-text>
+  </v-card>
+</template>
